requirements.txt
asgiref==3.5.0
Django==4.0.2
django-debug-toolbar==3.2.4
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.6
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
